implements new config format for GA

+++ config/metrics/adclicks.php
@@ -1,5 +1,5 @@
 <?php
 return [
     "id" => "adclicks",
-    "type" => "decimal"
+    "type" => "integer"
 ];
